fix mixins to be a bit more compatible
